+++ mobile/cmp/navigator/NavigatorModel.ts
@@ -8,9 +8,11 @@ import {HoistModel, RefreshMode, RenderMode, XH} from '@xh/hoist/core';
 import '@xh/hoist/mobile/register';
 import {action, bindable, makeObservable} from '@xh/hoist/mobx';
 import {ensureNotEmpty, ensureUniqueBy, throwIf, warnIf, mergeDeep} from '@xh/hoist/utils/js';
+import {wait} from '@xh/hoist/promise';
 import {find, isEqual, keys} from 'lodash';
-import {page} from './impl/Page';
+import {Swiper} from 'swiper/types';
 import {PageConfig, PageModel} from './PageModel';
+import {isScrollable} from './impl/Utils';
 
 export interface NavigatorConfig {
     /** Configs for PageModels, representing all supported pages within this Navigator/App. */
@@ -22,11 +24,14 @@ export interface NavigatorConfig {
      */
     track?: boolean;
 
-    /** True to enable 'swipe to go back' functionality. */
-    swipeToGoBack?: boolean;
-
     /** True to enable 'pull down to refresh' functionality. */
     pullDownToRefresh?: boolean;
+
+    /**
+     * Time (in milliseconds) for the transition between pages on route change.
+     * Defaults to 500.
+     */
+    transitionMs?: number;
 
     /**
      * Strategy for rendering pages. Can be set per-page via `PageModel.renderMode`.
@@ -42,7 +47,7 @@ export interface NavigatorConfig {
 }
 
 /**
- * Model for handling stack-based navigation between Onsen pages.
+ * Model for handling stack-based navigation between pages.
  * Provides support for routing based navigation.
  */
 export class NavigatorModel extends HoistModel {
@@ -52,31 +57,40 @@ export class NavigatorModel extends HoistModel {
     stack: PageModel[] = [];
 
     pages: PageConfig[] = [];
-
     track: boolean;
-    swipeToGoBack: boolean;
     pullDownToRefresh: boolean;
+    transitionMs: number;
     renderMode: RenderMode;
     refreshMode: RefreshMode;
 
-    private _navigator = null;
+    private _swiper: Swiper;
     private _callback: () => void;
-    private _prevKeyStack: string[];
 
     get activePageId(): string {
         return this.activePage?.id;
     }
 
     get activePage(): PageModel {
-        const {stack} = this;
-        return stack[stack.length - 1];
+        return this.stack[this.activePageIdx];
+    }
+
+    get activePageIdx(): number {
+        return this._swiper?.activeIndex ?? this.stack.length - 1;
+    }
+
+    get allowSlideNext(): boolean {
+        return this.activePageIdx < this.stack.length - 1;
+    }
+
+    get allowSlidePrev(): boolean {
+        return this.activePageIdx > 0;
     }
 
     constructor({
         pages,
         track = false,
-        swipeToGoBack = true,
         pullDownToRefresh = true,
+        transitionMs = 500,
         renderMode = 'lazy',
         refreshMode = 'onShowLazy'
     }: NavigatorConfig) {
@@ -92,19 +106,14 @@ export class NavigatorModel extends HoistModel {
 
         this.pages = pages;
         this.track = track;
-        this.swipeToGoBack = swipeToGoBack;
         this.pullDownToRefresh = pullDownToRefresh;
+        this.transitionMs = transitionMs;
         this.renderMode = renderMode;
         this.refreshMode = refreshMode;
 
         this.addReaction({
             track: () => XH.routerState,
             run: () => this.onRouteChange()
-        });
-
-        this.addReaction({
-            track: () => this.stack,
-            run: this.onStackChangeAsync
         });
 
         if (track) {
@@ -131,8 +140,50 @@ export class NavigatorModel extends HoistModel {
     //--------------------
     // Implementation
     //--------------------
-    private onRouteChange(init = null) {
-        if (!this._navigator || !XH.routerState) return;
+    /** @internal */
+    setSwiper(swiper: Swiper) {
+        if (this._swiper) return;
+        this._swiper = swiper;
+
+        this._swiper.on('transitionEnd', () => this.onPageChange());
+
+        // Cancel drag back if you are dragging a horizontally scrollable element.
+        this._swiper.on('touchStart', (s, event: TouchEvent) => {
+            this._swiper.allowTouchMove = !isScrollable(event, 'horizontal');
+        });
+
+        this._swiper.on('touchEnd', () => {
+            this._swiper.allowTouchMove = true;
+        });
+
+        this.onRouteChange(true);
+    }
+
+    /** @internal */
+    @action
+    onPageChange = () => {
+        // 1) Clear any pages after the active page. These can be left over from a back swipe.
+        this.stack = this.stack.slice(0, this._swiper.activeIndex + 1);
+
+        // 2) Sync route to match the current page stack
+        const newRouteName = this.stack.map(it => it.id).join('.'),
+            newRouteParams = mergeDeep({}, ...this.stack.map(it => it.props));
+
+        XH.navigate(newRouteName, newRouteParams);
+
+        // 3) Update state according to the active page and trigger optional callback
+        this.disableAppRefreshButton = this.activePage?.disableAppRefreshButton;
+        this._callback?.();
+        this._callback = null;
+
+        // 4) Remove finished shadow components. These are created during transitions,
+        // and remain overlaid on the page, preventing touch events from reaching the page.
+        // Presumably a bug in the Swiper library.
+        document.querySelectorAll('.swiper-slide-shadow-creative').forEach(e => e.remove());
+    };
+
+    private onRouteChange(init: boolean = false) {
+        if (!this._swiper || !XH.routerState) return;
 
         // Break the current route name into parts, and collect any params for each part.
         // Use meta.params to determine which params are associated with each route part.
@@ -154,7 +205,6 @@ export class NavigatorModel extends HoistModel {
 
         // Loop through the route parts, rebuilding the page stack to match.
         const stack = [];
-
         for (let i = 0; i < routeParts.length; i++) {
             const part = routeParts[i],
                 pageModelCfg = find(this.pages, {id: part.id});
@@ -176,85 +226,56 @@ export class NavigatorModel extends HoistModel {
                 return;
             }
 
-            const page = new PageModel({
-                navigatorModel: this,
-                ...mergeDeep({}, pageModelCfg, part)
-            });
-
-            stack.push(page);
+            // Re-use existing PageModels where possible
+            const existingPageModel = i < this.stack.length ? this.stack[i] : null;
+            if (
+                existingPageModel?.id === part.id &&
+                isEqual(existingPageModel?.props, part.props)
+            ) {
+                stack.push(existingPageModel);
+            } else {
+                stack.push(
+                    new PageModel({
+                        navigatorModel: this,
+                        ...mergeDeep({}, pageModelCfg, part)
+                    })
+                );
+            }
         }
 
-        this.stack = stack;
-    }
-
-    private async onStackChangeAsync() {
-        // Sync Onsen Navigator's pages with our stack
-        if (!this._navigator) return;
-        const {stack} = this,
-            keyStack = stack.map(it => it.key),
-            prevKeyStack = this._prevKeyStack || [],
-            backOnePage = isEqual(keyStack, prevKeyStack.slice(0, -1)),
-            forwardOnePage = isEqual(keyStack.slice(0, -1), prevKeyStack);
-
-        // Skip transition animation if the active page is going to be unmounted
-        let options;
-        if (this.activePage?.renderMode === 'unmountOnHide') {
-            options = {animation: 'none'};
+        // Immediately set the stack if this is the initial route change
+        if (init) {
+            this.stack = stack;
+            this._swiper.update();
+            this._swiper.activeIndex = this.stack.length - 1;
+            return;
         }
 
-        this._prevKeyStack = keyStack;
+        // Compare new stack to current stack to determine how to navigate
+        const {transitionMs} = this,
+            newKeyStack = stack.map(it => it.key),
+            currKeyStack = this.stack.map(it => it.key),
+            backOnePage = isEqual(newKeyStack, currKeyStack.slice(0, -1)),
+            forwardOnePage = isEqual(newKeyStack.slice(0, -1), currKeyStack);
 
         if (backOnePage) {
-            // If we have gone back one page in the same stack, we can safely pop() the page
-            return this._navigator.popPage(options);
-        } else if (forwardOnePage) {
-            // If we have gone forward one page in the same stack, we can safely push() the new page
-            return this._navigator.pushPage(stack[stack.length - 1], options);
+            // Don't update the stack yet. Instead, wait until after the animation has
+            // completed in onPageChange().
+            this._swiper.slidePrev(transitionMs);
         } else {
-            // Otherwise, we should reset the page stack
-            return this._navigator.resetPageStack(stack, {animation: 'none'});
+            // Otherwise, update the stack immediately and navigate to the new page.
+            this.stack = stack;
+            this._swiper.update();
+
+            // Wait for the new page to be rendered before sliding to it.
+            wait(1).then(() => {
+                if (forwardOnePage) {
+                    this._swiper.slideNext(transitionMs);
+                } else {
+                    // Jump instantly to the active page.
+                    this._swiper.slideTo(stack.length - 1, 0);
+                }
+            });
         }
     }
-
-    renderPage = (model, navigator) => {
-        const {init, key} = model;
-
-        // Note: We use the special 'init' object to obtain a reference to the
-        // navigator and to read the initial route.
-        if (init) {
-            if (!this._navigator) {
-                this._navigator = navigator;
-                this.onRouteChange(init);
-            }
-            return null;
-        }
-
-        // This is a workaround for an Onsen issue with resetPageStack(),
-        // which can result in transient duplicate pages in a stack. Having duplicate pages
-        // will cause React to throw with a duplicate key error. The error occurs
-        // when navigating from one page stack to another where the last page of
-        // the new stack is already present in the previous stack.
-        //
-        // For this workaround, we skip rendering the duplicate page (the one at the incorrect index).
-        //
-        // See https://github.com/OnsenUI/OnsenUI/issues/2682
-        const onsenNavPages = this._navigator.routes.filter(it => !it.init),
-            hasDupes = onsenNavPages.filter(it => it.key === key).length > 1;
-
-        if (hasDupes) {
-            const onsenIdx = onsenNavPages.indexOf(model),
-                ourIdx = this.stack.findIndex(it => it.key === key);
-
-            if (onsenIdx !== ourIdx) return null;
-        }
-
-        return page({model, key});
-    };
-
-    @action
-    onPageChange = () => {
-        this.disableAppRefreshButton = this.activePage?.disableAppRefreshButton;
-        this._callback?.();
-        this._callback = null;
-    };
 }